v3.2，首页金刚区科普功能页收集子页bug修复

+++ src/main/kotlin/com/example/kpappercutting/model/KnowledgeCollection.kt
@@ -1,3 +1,4 @@
+// src/main/kotlin/com/example/kpappercutting/model/KnowledgeCollection.kt
 package com.example.kpappercutting.model
 
 import jakarta.persistence.*
@@ -6,19 +7,14 @@ import java.time.LocalDateTime
 @Entity
 @Table(
     name = "knowledge_collections",
-    uniqueConstraints = [
-        UniqueConstraint(columnNames = ["user_id", "knowledge_id"])
-    ]
+    uniqueConstraints = [UniqueConstraint(columnNames = ["user_id", "knowledge_id"])]
 )
 data class KnowledgeCollection(
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0,
-
     @Column(name = "user_id", nullable = false)
     val userId: Long = 0,
-
     @Column(name = "knowledge_id", nullable = false)
-    val knowledgeId: Int = 0,
-
+    val knowledgeId: Long = 0, // 注意这里改成了 Long
     val createTime: LocalDateTime = LocalDateTime.now()
 )
